♻️ Refactor party size filter to component
